requirements.txt
requests>=2.31.0
pandas>=2.1.0
plotly>=5.18.0
kaleido>=0.2.1
jinja2>=3.1.0
weasyprint>=60.0
pyyaml>=6.0
numpy>=1.26.0
scipy>=1.11.0
akshare>=1.12.0
python-dotenv>=1.0.0
